refactor: move metric DTOs to data package
