+++ src/ai/backend/manager/api/gql_legacy/metric/user.py
@@ -9,11 +9,11 @@ from uuid import UUID
 import graphene
 
 from ai.backend.common.dto.clients.prometheus.request import QueryTimeRange
+from ai.backend.manager.data.metric.types import ContainerMetricOptionalLabel
 from ai.backend.manager.services.metric.actions.container import (
     ContainerMetricAction,
 )
 from ai.backend.manager.services.metric.types import (
-    ContainerMetricOptionalLabel,
     MetricQueryParameter,
 )
