+++ src/ai/backend/manager/services/metric/actions/container.py
@@ -3,13 +3,10 @@ from typing import override
 
 from ai.backend.common.data.permission.types import EntityType
 from ai.backend.common.dto.clients.prometheus.request import QueryTimeRange
+from ai.backend.manager.data.metric.types import ContainerMetricOptionalLabel, ContainerMetricResult
 from ai.backend.manager.services.metric.actions.base import (
     QueryMetricAction,
     QueryMetricActionResult,
-)
-from ai.backend.manager.services.metric.types import (
-    ContainerMetricOptionalLabel,
-    ContainerMetricResult,
 )
